[UPSTREAM?][sanitizers] Use usize instead of uptr for sizes and offsets
This will allow using the sanitizer runtime on the CHERI CPU where uptr is
a 128-bit capability type and usize is the 32/64-bit integer that we
actually want here. Also introduce a vaddr typedef for addresses.

+++ compiler-rt/lib/sanitizer_common/sanitizer_allocator_bytemap.h
@@ -41,7 +41,7 @@ class FlatByteMap {
 // to kSize2-byte arrays. The secondary arrays are mmaped on demand.
 // Each value is initially zero and can be set to something else only once.
 // Setting and getting values from multiple threads is safe w/o extra locking.
-template <u64 kSize1, u64 kSize2,
+template <usize kSize1, usize kSize2,
           typename AddressSpaceViewTy = LocalAddressSpaceView,
           class MapUnmapCallback = NoOpMapUnmapCallback>
 class TwoLevelByteMap {
